Included file
Moved all of the functions into a separete php file.
